docs(features): update access and ownership highlights for content management
- Add section detailing access rights for Posts, Pages, Media, and FavoriteClinics
- Clarify roles for Platform Staff and Patients regarding content creation and moderation
