+++ src/payload.config.ts
@@ -4,6 +4,7 @@ import { postgresAdapter } from '@payloadcms/db-postgres'
 import sharp from 'sharp'
 import path from 'path'
 import { buildConfig, PayloadRequest, PayloadHandler } from 'payload'
+import { seedPostHandler, seedGetHandler } from './endpoints/seed/seedEndpoint'
 import pino from 'pino'
 import { fileURLToPath } from 'url'
 
@@ -53,88 +54,8 @@ if (process.env.NODE_ENV === 'test') {
 
 export default buildConfig({
   endpoints: [
-    {
-      path: '/seed',
-      method: 'post',
-      handler: (async (req: PayloadRequest, res: any) => {
-        const payloadInstance = req.payload
-        const start = Date.now()
-        const type = (req.query.type as string) || 'baseline'
-        const reset = req.query.reset === '1'
-        try {
-          // Access control (platform staff only)
-          if (!req.user || (req.user as any).userType !== 'platform') {
-            return res.status(403).json({ error: 'Forbidden' })
-          }
-          if (process.env.NODE_ENV === 'production' && type !== 'baseline') {
-            return res.status(400).json({ error: 'Demo seeding disabled in production' })
-          }
-          const { runBaselineSeeds } = await import('./endpoints/seed/baseline')
-          const { runDemoSeeds } = await import('./endpoints/seed/demo')
-          if (type === 'baseline') {
-            const results = await runBaselineSeeds(payloadInstance)
-            const summary = {
-              type,
-              reset,
-              status: 'ok',
-              baselineFailed: false,
-              startedAt: new Date(start).toISOString(),
-              finishedAt: new Date().toISOString(),
-              durationMs: Date.now() - start,
-              totals: {
-                created: results.reduce((a: number, r: any) => a + r.created, 0),
-                updated: results.reduce((a: number, r: any) => a + r.updated, 0),
-              },
-              units: results,
-            }
-            ;(global as any).__lastSeedRun = summary
-            return res.json(summary)
-          } else if (type === 'demo') {
-            const outcome = await runDemoSeeds(payloadInstance, { reset })
-            const created = outcome.units.reduce((a, r) => a + r.created, 0)
-            const updated = outcome.units.reduce((a, r) => a + r.updated, 0)
-            let status: 'ok' | 'partial' | 'failed'
-            if (outcome.units.length === 0 && outcome.partialFailures.length > 0) {
-              status = 'failed'
-            } else if (outcome.partialFailures.length > 0) {
-              status = 'partial'
-            } else {
-              status = 'ok'
-            }
-            const summary = {
-              type,
-              reset,
-              status,
-              baselineFailed: false,
-              startedAt: new Date(start).toISOString(),
-              finishedAt: new Date().toISOString(),
-              durationMs: Date.now() - start,
-              totals: { created, updated },
-              units: outcome.units,
-              partialFailures: outcome.partialFailures,
-            }
-            ;(global as any).__lastSeedRun = summary
-            const httpStatus = status === 'failed' ? 500 : 200
-            return res.status(httpStatus).json(summary)
-          } else {
-            return res.status(400).json({ error: 'Invalid type parameter' })
-          }
-        } catch (e: any) {
-          payloadInstance.logger.error(`Seed endpoint error: ${e.message}`)
-          return res.status(500).json({ error: 'Seed failed', detail: e.message })
-        }
-      }) as PayloadHandler,
-    },
-    {
-      path: '/seed',
-      method: 'get',
-      handler: (async (req: PayloadRequest, res: any) => {
-        if (!req.user || (req.user as any).userType !== 'platform') {
-          return res.status(403).json({ error: 'Forbidden' })
-        }
-        return res.json((global as any).__lastSeedRun || { message: 'No seed run yet' })
-      }) as PayloadHandler,
-    },
+    { path: '/seed', method: 'post', handler: seedPostHandler as PayloadHandler },
+    { path: '/seed', method: 'get', handler: seedGetHandler as PayloadHandler },
   ],
   admin: {
     components: {
